# Experiment with passing in half of the training samples in finalizeDictionary